Adjusted calculator class further. Removed history-related tests from calculator_test and added them to calculations_history_test. Formatted calculator_test file to be cleaner and work with list args instead of "value_a and value_b"
